feat : 이미지 완료, SizeEntity 추가 완료, QAuctionRepo/QAwardRepo 완료, QProductRepo 경매 수 orderBy 수정 예정

+++ biday-back/src/main/java/shop/biday/model/domain/AwardModel.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
 import shop.biday.model.dto.AuctionDto;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -19,6 +20,6 @@ public class AwardModel {
     private AuctionDto auction;
     private String user;
     private LocalDateTime bidedAt;
-    private Long currentBid;
+    private BigInteger currentBid;
     private int count;
 }
